requirements.txt
django==4.2.*
environs==9.5.0
python-telegram-bot==13.15
django-admin-sortable2==2.1.8
